% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis_functions.R
\name{basis_functions}
\alias{basis_functions}
\alias{fourier_2d}
\alias{polynomial_2d}
\title{Creation of basis functions}
\usage{
fourier_2d(max_freqs, lattice_dim)

polynomial_2d(poly_deg, lattice_dim)
}
\arguments{
\item{max_freqs}{A length 2 numeric vector with maximum frequencies considered
(x-axis and y-axis direction, respectively).}

\item{lattice_dim}{A length 2 numeric vector with lattice dimensions (N,M)
to be used.}

\item{poly_deg}{A length 2 numeric vector with degrees of the bivariate
polynomial to be considered.}
}
\description{
Creation of basis functions
}
\author{
Victor Freguglia
}
