% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{bold5000}
\alias{bold5000}
\title{BOLD5000 neuroimaging data}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 176 rows and 256 columns.
}
\usage{
bold5000
}
\description{
An image extracted from the "BOLD5000" open dataset. It was read from
the file in path \verb{BOLD5000/DS001499/SUB-CSI2/SES-16/ANAT/SUB-CSI2_SES-16_T1W.NII.GZ},
available at the OpenNeuro platform (https://openneuro.org/datasets/ds001499/versions/1.3.0).
}
\details{
The file was read using the \code{oro.nifti} package and the image was extracted from the
matrix in slice 160.
}
\references{
Chang, N., Pyles, J. A., Marcus, A., Gupta, A., Tarr, M. J., & Aminoff, E. M. (2019).
BOLD5000, a public fMRI dataset while viewing 5000 visual images. Scientific data, 6(1), 1-18.
}
\seealso{
A paper with detailed description of the package can be found at
\doi{10.18637/jss.v101.i08}.
}
\keyword{datasets}
