% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{data_examples}
\alias{data_examples}
\alias{field1}
\alias{hfield1}
\title{Example Data}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 150 rows and 150 columns.

An object of class \code{matrix} (inherits from \code{array}) with 150 rows and 150 columns.
}
\usage{
field1

hfield1
}
\description{
\code{mrf2d} contains a set of simulated fields to illustrate its
usage.
\describe{
\item{field1}{A binary field sampled from a sparse interaction structure:
\code{mrfi(1) + c(4,4)}}
\item{hfield1}{A continuous valued field, obtained by Gaussian mixture driven
by \code{field1}.}
}
}
\author{
Victor Freguglia
}
\keyword{datasets}
