% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lplot.R
\name{dplot}
\alias{dplot}
\alias{cplot}
\title{Elegant image plots based on matrices}
\usage{
dplot(Z, legend = FALSE)

cplot(Y, legend = TRUE)
}
\arguments{
\item{Z}{A \code{matrix} object with integers only.}

\item{Y}{A \code{matrix} object with continuous values.}
}
\value{
a \code{ggplot} object.
}
\description{
An alternative to base R's \code{image} function using \code{ggplot2}.
Use \code{dplot} for discrete valued matrices and \code{cplot} for continuous value.
}
\author{
Victor Freguglia
}
