% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ghm.R
\name{fit_ghm}
\alias{fit_ghm}
\title{EM for Gaussian Hidden Markov field}
\usage{
fit_ghm(Y, mrfi, theta, fixed_fn = list(), equal_vars = TRUE,
  init_mus = NULL, init_sigmas = NULL, maxiter = 100,
  max_dist = 10^-3, icm_cycles = 6, verbose = TRUE)
}
\arguments{
\item{Y}{A matrix of observed pixel values.}

\item{mrfi}{A \code{\link[=mrfi-class]{mrfi}} object representing the
interaction structure.}

\item{theta}{A 3-dimensional array describing potentials. Slices represent
interacting positions, rows represent pixel values and columns represent
neighbor values. As an example: \code{theta[1,3,2]} has the potential pairs of
values 0,2 in the second relative position of \code{mrfi}.}

\item{fixed_fn}{A list of functions \code{fn(x,y)} to be considered as a fixed effect.}

\item{equal_vars}{\code{logical} indicating if the mixture model has the same
variances in all components}

\item{maxiter}{The maximum number of iterations allowed. Defaults to 100.}

\item{icm_cycles}{Number of steps used in the Iterated Conditional Modes
algorithm executed in each interaction. Defaults to 6.}
}
\value{
A list containing some stuff to be documented.
}
\description{
Em algorithm fitting of Gaussian mixture driven by a hidden
Markov random field with known parameters.
}
\details{
Add details about the estimation procedure.
}
\author{
Victor Freguglia
}
