% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudo_likelihood.R
\name{fit_pl}
\alias{fit_pl}
\title{Maximum Pseudo-likelihood fitting of MRFs on 2d lattices.}
\usage{
fit_pl(Z, mrfi, family = "onepar", init = 0, optim_args = list(method
  = "BFGS"), return_optim = FALSE)
}
\arguments{
\item{Z}{A \code{matrix} object containing the observed MRF.}

\item{mrfi}{A \code{\link[=mrfi-class]{mrfi}} object representing the
interaction structure.}

\item{family}{The family of restrictions to potentials. It can be:
\itemize{
\item \code{"onepar"}: Models with a single parameter. It is a single value for any
equal valued pairs and the same value with opposite signal for different
valued pairs.
}}

\item{init}{The initial value to be used in the optimization. It can be:
\itemize{
\item An \code{array} with vaid potential values according to \code{family}.
\item A \code{numeric} of length 1 if \code{family} is \code{onepar}.
\item \code{0}. If set to \code{0} an array with `0`` in all entries is created.
}}

\item{optim_args}{Additional parameters passed in \code{optim()} function call.}

\item{return_optim}{\code{logical} indicating whether information from the
\code{optim()} call are returned.}
}
\value{
A \code{list} object with elements:
\itemize{
\item \code{theta}: The array of estimated potential values.
\item \code{value}: The optimal pseudo-likelihood value.
\item \code{opt.xxx}(if \code{return_optim} is \code{TRUE}): Information returned by the
\code{optim()} function used for the optimization.
}
}
\description{
Parameter estimation for Markov random fields via
Pseudo-Likelihood function optimization. See
\code{\link[=pl_mrf2d]{pl_mrf2d}} for more information on the
Pseudo-Likelihood function.
}
\author{
Victor Freguglia
}
