% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmrfout.R
\name{hmrfout}
\alias{hmrfout}
\alias{print.hmrfout}
\alias{summary.hmrfout}
\alias{plot.hmrfout}
\title{MRF fitting functions output}
\usage{
\method{print}{hmrfout}(x, ...)

\method{summary}{hmrfout}(object, ...)

\method{plot}{hmrfout}(x, ...)
}
\arguments{
\item{x}{a \code{hmrfout} object.}

\item{...}{other arguments not used by this method.}

\item{object}{a \code{hmrfout} object.}
}
\description{
MRF fitting functions output
}
