% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family.R
\name{mrf2d-family}
\alias{mrf2d-family}
\title{Parameter restriction families}
\description{
Different parameter restrictions can be included in estimation processes
to make sure \code{mrf2d} can successfully include a wide range of model types in
its inference functions.

For model identifiability, at least one linear restriction is necessary.
\code{mrf2d} always assume \eqn{\theta_{0,0,r} = 0} for all relative positions
\eqn{r}.

Additionally, each family of restrictions may introduce other restrictions:
}
\section{'onepar'}{

This family assumes the model is defined by a single parameter
by adding the restriction

\deqn{\theta_{a,b,r} = \phi * 1(a != b).}

Here \eqn{1()} denotes de indicator function. In words, the parameter must
be the same value for any pair with different values and 0 for any
equal-valued pair.
}

\section{'oneeach'}{

Similar to \code{'onepar'}, parameters are 0 for equal-valued pairs and a
constant for pairs with different values, but the constant may differ
between different relative positions \eqn{r}:

\deqn{\theta{a,b,r} = \phi_r * 1(a != b).}
}

\section{'absdif'}{

All parameters \eqn{\theta_{a,b,r}} with the same absolute difference
between \eqn{a} and \eqn{b} must be equal within each relative position
\eqn{r}. (Note that \code{'absdif'} is equal to \code{'oneeach'} for binary images).

\deqn{\theta_{a,b,r} = \sum_d \phi_{d,r} * 1(|a-b| == d)}
}

\section{'dif'}{

The same as \code{'absdif'}, but parameters may differ between positive and
negative differences.

\deqn{\theta_{a,b,r} = \sum_d \phi_{d,r} * 1(a-b == d)}
}

\section{'free'}{

No additional restriction, all parameters other than \eqn{\theta_{0,0,r}}
vary freely.
}

\seealso{
\code{vignette("mrf2d-family", package = "mrf2d")}

A paper with detailed description of the package can be found at
\url{https://arxiv.org/abs/2006.00383}
}
\author{
Victor Freguglia
}
