% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrf2d.R
\docType{package}
\name{mrf2d-package}
\alias{mrf2d}
\alias{mrf2d-package}
\title{mrf2d: Markov Random Field Models for Image Analysis}
\description{
\code{mrf2d} contains tools for Markov Random Field models on
two-dimensional lattices.

To learn more about \code{mrf2d}, read the paper at
\doi{10.18637/jss.v101.i08}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Freguglia/mrf2d}
  \item Report bugs at \url{https://github.com/Freguglia/mrf2d/issues}
}

}
\author{
\strong{Maintainer}: Victor Freguglia \email{victorfreguglia@gmail.com} (\href{https://orcid.org/0000-0002-6189-4453}{ORCID})

}
\keyword{internal}
