% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrfi_class.R
\docType{class}
\name{mrfi-class}
\alias{mrfi-class}
\title{The \code{mrfi} class}
\description{
A representation of the interaction structure for a spatially-stationary
Markov Random Field.
}
\section{Slots}{

\describe{
\item{\code{Rmat}}{A 2-column \code{matrix} where each row represents a relative position of
interaction.}

\item{\code{n_neis}}{The number of intracting positions.}
}}

