% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrfi_methods.R
\name{mrfi-operations}
\alias{mrfi-operations}
\alias{-,mrfi,numeric-method}
\alias{+,mrfi,mrfi-method}
\title{Set operations for \code{mrfi} objects}
\usage{
\S4method{+}{mrfi,numeric}(e1, e2)

\S4method{-}{mrfi,numeric}(e1, e2)

\S4method{+}{mrfi,mrfi}(e1, e2)
}
\arguments{
\item{e1}{A \code{mrfi} object.}

\item{e2}{Either a second \code{mrfi} object or a length 2 \code{numeric} with the new
relative position to include (\verb{+}) or remove (\verb{-}).}
}
\description{
Set operations for \code{mrfi} objects
}
\examples{
mrfi(1) + c(2,0)
mrfi(1) - c(1,0)
mrfi(1) + mrfi(0, positions = list(c(2,0)))
}
