% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrfi_methods.R
\name{mrfi-subsetting}
\alias{mrfi-subsetting}
\alias{[,mrfi,numeric,missing-method}
\title{Subsetting \code{mrfi} objects}
\usage{
\S4method{[[}{mrfi,numeric,missing}(x, i)

\S4method{[}{mrfi,numeric,missing}(x, i)
}
\arguments{
\item{i}{vector of indexes to extract interacting positions.}
}
\value{
\verb{[[}: converts to list and subsets it.

\verb{[}: subsets the \code{mrfi} object and returns another \code{mrfi} object.

\verb{+}: computes the union of the interaction structure in a \code{mrfi} object with
a \code{numeric} representing an additional position to include or another \code{mrfi}
object.
}
\description{
Subsetting \code{mrfi} objects
}
