% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrfi_class.R
\name{mrfi}
\alias{mrfi}
\title{Creation of \code{\link[=mrfi-class]{mrfi}} objects.}
\usage{
mrfi(max_norm = 1, positions = NULL, norm_type = "1")
}
\arguments{
\item{max_norm}{a \code{numeric} value. All points with norm \eqn{\le} \code{max_dist} are included.}

\item{positions}{a \code{list} of \code{numeric} vectors of length 2. Each vector corresponds to a relative position included.}

\item{norm_type}{a \code{character} indicating the norm type used. Possible values are "m", "1", "2", etc. See \code{\link[=norm]{norm}} for details.}
}
\value{
A \code{\link[=mrfi-class]{mrfi}} object.
}
\description{
Creation of \code{\link[=mrfi-class]{mrfi}} objects.
}
