% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrfout.R
\name{mrfout}
\alias{mrfout}
\alias{print.mrfout}
\title{MRF fitting functions output}
\usage{
\method{print}{mrfout}(x, ...)
}
\arguments{
\item{x}{a \code{mrfout} object.}

\item{...}{other arguments not used by this method}
}
\description{
MRF fitting functions output
}
