% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrfi_class.R
\docType{methods}
\name{plot.mrfi}
\alias{plot.mrfi}
\alias{mrfi-plot}
\title{Plotting of \code{mrfi} objects.}
\usage{
\S4method{plot}{mrfi,missing}(x, no_axis = FALSE, no_text = FALSE)
}
\arguments{
\item{x}{A \code{\link[=mrfi-class]{mrfi}} object.}

\item{no_axis}{\code{logical} value indicating whether the axis and grid lines are present.}

\item{no_text}{\code{logical} value indicating whether interaction types should be included.}
}
\description{
Plotting of \code{mrfi} objects.
}
