% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmrf.R
\name{rmrf}
\alias{rmrf}
\title{Sampling of Markov Random Fields}
\usage{
rmrf(init_Z, mrfi, theta, steps = 10, method = "gibbs")
}
\arguments{
\item{init_Z}{A \code{matrix} object with the initial field configuration. Its values
must be integers in 0,...,C.}

\item{mrfi}{A \code{\link[=mrfi-class]{mrfi}} object representing the
interaction structure.}

\item{theta}{A 3-dimensional array describing potentials. Slices represent
interacting positions, rows represent pixel values and columns represent
neighbor values. As an example: \code{theta[1,3,2]} has the potential pairs of
values 0,2 in the second relative position of \code{mrfi}.}

\item{steps}{The number of complete (all pixels) updates to be done.}

\item{method}{Method used to perform pixel-wise updates.
\itemize{
\item \code{'gibbs'}: Gibbs Sampler method. Samples from conditional distribution.
}}
}
\description{
Performs pixel-wise updates based on conditional distributions
to sample from a Markov random field. The order pixels are updated is
randomized in each step.
}
