% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmrf.R
\name{rmrf2d}
\alias{rmrf2d}
\title{Sampling of Markov Random Fields on 2d lattices}
\usage{
rmrf2d(init_Z, mrfi, theta, cycles = 10, method = "gibbs")
}
\arguments{
\item{init_Z}{One of two options:
\itemize{
\item A \code{matrix} object with the initial field configuration. Its
valuesmust be integers in \code{0,...,C}.
\item A length 2 \code{numeric} vector with the lattice dimensions.
}}

\item{mrfi}{A \code{\link[=mrfi-class]{mrfi}} object representing the
interaction structure.}

\item{theta}{A 3-dimensional array describing potentials. Slices represent
interacting positions, rows represent pixel values and columns represent
neighbor values. As an example: \code{theta[1,3,2]} has the potential pairs of
values 0,2 in the second relative position of \code{mrfi}.}

\item{cycles}{The number of complete (all pixels) updates to be done.}

\item{method}{Method used to perform pixel-wise updates.
\itemize{
\item \code{'gibbs'}: Gibbs Sampler method. Samples from conditional distribution.
}}
}
\description{
Performs pixelwise updates based on conditional distributions
to sample from a Markov random field.
}
\details{
This function implements a Gibbs Sampling scheme to sample from
a Markov random field by iteratively sampling pixel values from the
conditional distribution
\deqn{P(Z_i | Z_{{N}_i}).}

A cycle means exactly one update to each pixel. The order pixels are
sampled is randomized within each cycle.

If \code{init_Z} is passed as a length 2 vector with lattice dimensions, the
initial field is sampled from independent discrete uniform distributions in
\code{0,...,C}. The value of C is obtained from the number of rows/columns of
\code{theta}.
}
\author{
Victor Freguglia
}
