% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_icm.R
\name{fit_icm}
\alias{fit_icm}
\title{Iterated Conditional Modes algorithm for image restoration}
\usage{
fit_icm(init_Z, mrfi, theta, error_prob, cycles = 10)
}
\arguments{
\item{mrfi}{A \code{\link[=mrfi-class]{mrfi}} object representing the
interaction structure.}

\item{theta}{A 3-dimensional array describing potentials. Slices represent
interacting positions, rows represent pixel values and columns represent
neighbor values. As an example: \code{theta[1,3,2]} has the potential pairs of
values 0,2 in the second relative position of \code{mrfi}.}

\item{error_prob}{The probability a pixel's value is corrupted.}

\item{n_cycles}{The number of times each pixel is updated.}
}
\value{
A \code{matrix} with the resulting configuration.
}
\description{
Computes maximum probability configuration for a Markov random
with corrupted pixel values.
}
\author{
Victor Freguglia
}
