% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochastic_approximation.R
\name{fit_sa}
\alias{fit_sa}
\title{Stochastic Approximation algorithm for MRFs on 2d lattices}
\usage{
fit_sa(Z, mrfi, family = "onepar", gamma_seq, init = 0, steps = 5)
}
\arguments{
\item{Z}{A \code{matrix} object containing the observed MRF.}

\item{mrfi}{A \code{\link[=mrfi-class]{mrfi}} object representing the
interaction structure.}

\item{family}{The family of parameter restrictions to potentials. Families
are:
\code{'onepar'}, \code{'oneeach'}, \code{'absdif'}, \code{'dif'} or \code{'free'}.
See \code{\link[=mrf2d-family]{mrf2d-familiy}}.}

\item{gamma_seq}{A \code{numeric} vector with the sequence of constants \eqn{\gamma_t}}

\item{init}{The initial value to be used in the optimization. It can be:
\itemize{
\item A valid \code{array} of parameter values according to \code{family}.
\item \code{0}. If set to \code{0} an array with `0`` in all entries is created.
}}

\item{steps}{Number of Gibbs Sampler updates between samples.}
}
\value{
A \code{list} object with the following elements:
\itemize{
\item \code{theta}: The estimated \code{array} of potentials.
\item \code{metrics}: A \code{data.frame} containing the the euclidean distance between
the sufficient statics computed for \code{Z} and the current sample.
}
}
\description{
Parameter estimation for Markov Random Fieds via Stochastic
Approximation algorithm.
}
\author{
Victor Freguglia
}
