% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_sa.R
\name{fit_sa}
\alias{fit_sa}
\title{Stochastic Approximation algorithm for MRFs on 2d lattices}
\usage{
fit_sa(Z, mrfi, family = "onepar", gamma_seq, init = 0, cycles = 5,
  refresh_each = length(gamma_seq) + 1, refresh_cycles = 60)
}
\arguments{
\item{Z}{A \code{matrix} object containing the observed MRF.}

\item{mrfi}{A \code{\link[=mrfi-class]{mrfi}} object representing the
interaction structure.}

\item{family}{The family of parameter restrictions to potentials. Families
are:
\code{'onepar'}, \code{'oneeach'}, \code{'absdif'}, \code{'dif'} or \code{'free'}.
See \code{\link[=mrf2d-family]{mrf2d-familiy}}.}

\item{gamma_seq}{A \code{numeric} vector with the sequence of constants
used in each step \eqn{\gamma_t}.}

\item{init}{The initial value to be used in the optimization. It can be:
\itemize{
\item A valid \code{array} of parameter values according to \code{family}.
\item \code{0}. If set to \code{0} an array with `0`` in all entries is created.
}}

\item{cycles}{The number of updates to be done (for each each pixel).}

\item{refresh_each}{An integer with the number of iterations taken before a
complete refresh (restart from a random state). This prevents the sample from
being stuck in a mode for too long. Defaults to \code{length(gamma_seq) + 1} (no
refresh happens).}

\item{refresh_cycles}{An iteger indicating how many Gibbs Sampler cycles are
performed when a refresh happens. Larger is usually better, but slower.}
}
\value{
A \code{list} object with the following elements:
\itemize{
\item \code{theta}: The estimated \code{array} of potentials.
\item \code{metrics}: A \code{data.frame} containing the the euclidean distance between
the sufficient statics computed for \code{Z} and the current sample.
}
}
\description{
Estimates the parameters of a MRF by successively sampling from
the a parameter and updating it by comparing the sufficient statistics
of the sampled field and the reference field.

This method aims to find the parameter value where the gradient of the
likelihood function is equal to zero.
}
\details{
\deqn{\theta_{t+1} = \theta_t - \gamma_t(T(Z_t) - T(Z)),}
where \eqn{T(Z)} is the sufficient statistics for the reference field,
\eqn{T(Z_t)} is the sufficient statistics for a field sampled from
\eqn{\theta_t}.
}
\note{
Stochastic Approximation is called "Controllable Simulated Annealing" in
some references.

Examples where Stochastic Approximation is used with MRFs are
\insertCite{gimel_sa}{mrf2d}, \insertCite{CR}{mrf2d}.
}
\examples{
\dontrun{
set.seed(2)
}

}
\references{
\insertRef{wiki_sa}{mrf2d}

\insertAllCited{}
}
\author{
Victor Freguglia
}
