% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrfi_class.R
\docType{class}
\name{mrfi-class}
\alias{mrfi-class}
\title{The mrfi class}
\description{
A representation for the interaction structure of a spatially-stationary
Markov Random Field.
}
\section{Slots}{

\describe{
\item{\code{Rmat}}{A 3-column matrix where each row represents a relative position of
interaction. First and second columns represent the position and third column
the interaction type.}

\item{\code{n_neis}}{The number of intracting positions.}

\item{\code{n_types}}{The number of types of interaction.}
}}

