% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrfi_methods.R
\name{[[,mrfi,numeric,missing-method}
\alias{[[,mrfi,numeric,missing-method}
\alias{[,mrfi,numeric,missing-method}
\title{Subsetting \code{mrfi} objects}
\usage{
\S4method{[[}{mrfi,numeric,missing}(x, i)

\S4method{[}{mrfi,numeric,missing}(x, i)
}
\arguments{
\item{x}{\code{mrfi} object.}

\item{i}{vector of indexes to extract interacting positions.}
}
\value{
\code{[[}: converts to list and subsets it.

\code{[}: subsets the \code{mrfi} object and returns another \code{mrfi} object.

\code{+}: computes the union of the interaction structure in a \code{mrfi} object with
a \code{numeric} representing an additional position to include or another \code{mrfi}
object.
}
\description{
Subsetting \code{mrfi} objects
}
