% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrfi_class.R
\docType{methods}
\name{plot.mrfi}
\alias{plot.mrfi}
\alias{mrfi-plot}
\title{Plotting of \code{mrfi} objects.}
\usage{
\S4method{plot}{mrfi,missing}(x, no_axis = FALSE)
}
\arguments{
\item{x}{A \code{\link[=mrfi-class]{mrfi}} object.}

\item{no_axis}{\code{logical} value indicating whether the axis and grid lines are present.}
}
\description{
Plots a visual description of the interaction structure
described in a \code{mrfi} object. The black tile represents a reference pixel
and gray tiles are shown in relative positions where pixels are
conditionally dependent.
}
